requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Data Collection
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
praw==7.7.1
feedparser==6.0.10

# Time Series & Sentiment Analysis
statsmodels==0.14.0
vaderSentiment==3.3.2
textblob==0.17.1

# Data Storage
sqlalchemy==2.0.20

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

